+++ coolstrap-core/spec/middleman/helpers_spec.rb
@@ -1,0 +1,16 @@
+require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
+
+require "middleman"
+require "coolstrap-core/helpers"
+
+describe "Middleman Helpers" do
+  
+  before(:each) do
+    
+  end
+  
+  it "should return the base location" do
+    
+  end
+  
+end