+++ coolstrap-core/spec/spec_helper.rb
@@ -1,0 +1,6 @@
+$LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib', 'coolstrap-core'))
+$LOAD_PATH.unshift(File.dirname(__FILE__))
+require 'rspec'
+require 'coolstrap-core'
+require 'stringio'
+#require 'config'