+++ coolstrap-gen/lib/coolstrap-gen/builder/middleman.rb
@@ -1,0 +1,14 @@
+module Coolstrap::Gen
+  module Builder
+    class Middleman
+      class << self
+        include ::Coolstrap::Gen::Utils
+        
+        def build
+          system "middleman build"
+        end
+
+      end
+    end
+  end
+end