+++ coolstrap/lib/coolstrap.rb
@@ -1,0 +1,9 @@
+# Depends on both libraries
+require "coolstrap-core"
+require "coolstrap-gen"
+
+# Make the VERSION string available
+require "coolstrap/version"
+
+module Coolstrap
+end