+++ source/examples/components/app/events.js
@@ -1,8 +1,0 @@
-App.Events = (function(cool) {
-
-  $(document).ready(function() {
- 
-  });
- 
-
-})(COOLSTRAP);