+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Coolstrap Changelog
 
+
++ 16/jul/2012
+  + multi gem layout support.
+  + merging coolstrap-generator into coolstrap, and split coolstrap in coolstrap-gen & coolstrap-core.
+  + moving helpers into coolstrap.
+  + moving coolstrap assets into coolstrap-core.
 + 7/jul/2012
   change folder structure to be more like asset pipelines